requirements.txt
streamlit
pandas
easyocr
Pillow